📦 chore(package.json): add "cube-notation-normalizer" and "scrambo" as dependencies 📝 docs(makeScramble.ts): add function to create a normalized scramble for FMC competition
The "cube-notation-normalizer" and "scrambo" packages were added as dependencies in the package.json file. This is to support the creation of a normalized scramble for the FMC (Fewest Moves Challenge) competition.

A new file "makeScramble.ts" was created in the "src/services" directory. This file contains a function called "createFmcScrambleToNormalScramble" which generates a scramble using the "scrambo" package and normalizes it using the "cube-notation-normalizer" package. The normalized scramble is then returned along with its inverse.

This function is used to ensure that the generated scramble follows the rules and conventions of the FMC competition.